LayoutTests/http/tests/security/contentSecurityPolicy/style-blocked-expected.txt

CONSOLE MESSAGE: Refused to load style from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/blue.css' because of Content-Security-Policy.

PASS
